Rename customer to actor because they can also be administrators.
Add administrators with their associated powers.  A single actor can
have both account privileges and admin powers.

Create an admin API to create new offers.

+++ src/entities/trading.rs
@@ -26,7 +26,7 @@ impl OrderState {
 #[derive(Clone)]
 pub struct Order {
     pub order_id: i64,
-    pub account_id: i64,
+    pub account_id: i32,
     pub order_number: i32,
     pub ext_order_id: String,
     pub client_order_id: String,
