+++ src/persistence/account.rs
@@ -18,7 +18,7 @@ impl<'b> DaoTransaction<'b> {
         Ok(account)
     }
 
-    pub async fn get_account(&self, account_id: i64) -> Result<Option<Account>, DaoError> {
+    pub async fn get_account(&self, account_id: i32) -> Result<Option<Account>, DaoError> {
         let accounts_map = match self.get_accounts(vec![account_id]).await {
             Ok(accounts_map) => accounts_map,
             Err(db_error) => return Err(db_error)
@@ -29,7 +29,7 @@ impl<'b> DaoTransaction<'b> {
         })
     }
 
-    pub async fn get_accounts(&self, account_ids: Vec<i64>) -> Result<HashMap<i64, Account>, DaoError> {
+    pub async fn get_accounts(&self, account_ids: Vec<i32>) -> Result<HashMap<i32, Account>, DaoError> {
         let mut query_string: String = "".to_owned();
         query_string.push_str(ACCOUNT_QUERY);
         query_string.push_str(" WHERE account.accountId = ANY ($1)");
